Reorder synthesized members to come after all non-synthesized members in MRO
Summary:
Synthesized members are those inherited from synthesized ancestors--any ancestor which comes from a [require-extends or require-implements relationship](https://hhvm.com/blog/9581/trait-and-interface-requirements-in-hack). We want the members defined in the synthesized ancestor to be present in the trait which requires it so that we can typecheck uses of those members within the trait, but we don't want to override more specific, non-synthesized members with more general synthesized ones. For example, consider this file (equivalent to the typechecker test trait_require_extends.php):

```
class Base {
  public function foo(): ?int {
    return null;
  }
}

class Child extends Base {
  public function foo(): int {
    return 0;
  }
}

trait T1 {
  require extends Base;
  public function bar(): ?int {
    return $this->foo();
  }
}

class Grandchild extends Child {
  use T1;
}
```

In order to typecheck `T1::bar`, we must verify that T1 has a method named `foo`, and it is the require-extends relationship that guarantees this. T1's methods table includes the method `foo`, and it is marked as synthesized, since it came from T1's synthesized ancestor Base.

Class Grandchild extends Child, which overrides `foo` to return `int` rather than `?int`. In Grandchild's methods table, we would expect to see a non-synthesized definition of `foo` which returns `int`.

A naive usage of the linearization would not produce this result. Before this diff, the linearization of Grandchild is:

```
[Grandchild, T1, Base(synthesized), Child, Base]
```

So if we attempt to look up the method `foo` and stop at the first concrete implementation, we will get the definition from Base, which is not the one we want.

Historically, we dealt with this problem by considering the synthesized flag during the folding of inherited member tables. [Here is where we do so today](diffusion/FBS/browse/master/fbcode/hphp/hack/src/decl/decl_inherit.ml;f4c8283c966de8eb851cbb5cd9e86ea5897f0d8c$60)--we prefer an ancestor signature over one that might override it if the ancestor signature is not synthesized and the overriding signature is synthesized.

Unfortunately, it is not sufficient to do the same when looking up members using the linearization. The above example will work--Child's definition will be the first concrete, non-synthesized definition we encounter while traversing the linearization--but there are other cases that don't. Suppose we add this trait T2 to the above example:

```
trait T2 {
  require extends Grandchild;

  public function test(): void {
    takes_int($this->foo());
  }
}
```

The linearization of T2 is this:

```
[T2, Grandchild(synthesized), T1(synthesized), Base(synthesized), Child(synthesized)]
```

Since the *entire* hierarchy is synthesized, T2 will inherit Base's definition rather than Child's, and we will emit a typechecker error at the `takes_int` invocation (because we believe `T2::foo` returns `?int`, as Base specifies).

Clearly, it is not sufficient to simply mark ancestors as synthesized. Perhaps instead of a flag, we could use an integer *degree* of synthesized-ness... but this sounds like a terrible hack (and it is).

Let's take a step back. Ideally, the linearization should be a straightforward member resolution order--we should be able to iterate over it and take the first implementation we find (in practice, we can't quite do this--we need to skip over abstract definitions if a concrete definition exists higher up--but the less of this special handling we do during lookup, the better). If some ancestors take priority over others, they should come first in the MRO. Can we de-prioritize synthesized ancestors simply by placing them at the end of the linearization?

This seems to work, both for the original example, and after the addition of T2:

```
Linearization for class Base:
[Base]

Linearization for class Child:
[Child, Base]

Linearization for class T1:
[T1, Base(synthesized)]

Linearization for class Grandchild:
[Grandchild, T1, Child, Base, Base(synthesized)]

Linearization for class T2:
[T2, Grandchild(synthesized), T1(synthesized), Child(synthesized), Base(synthesized)]
```

Now, the first concrete definition of `foo` in T2's linearization is the one from Child, because it preceded Base(synthesized) in Grandchild's linearization.

Reviewed By: arxanas

Differential Revision: D13579467

fbshipit-source-id: 1b21d6991c54c54352bb54c6391d266c2b9c18f7