Add Typing_class_member_cache
Summary:
While prototyping lazy lookup for ancestors and methods using the linearization, I quickly found that a naive, cache-free approach caused much too great a performance regression (as expected). We clearly need some means of caching lookups of methods and ancestors (as well as other class members). Since we already have an OCaml-heap cache of folded class types (Typing_classes_heap), this seems like an appropriate place to do this caching. All we need is some caching collection type which takes a sequence (in the case of methods, the sequence of all inherited methods derived from the linearization), lazily advances the sequence as necessary to answer lookup queries, and answers those queries from a cache whenever possible. This would guarantee that we traverse the linearization only once for each of these caches.

This diff adds such a cache type, `Typing_class_member_cache.t`, which provides a memoization cache for any `(string * 'a) Sequence.t` where:

1. It is desirable to look up elements in the sequence by their `string` key
2. Conflicts between multiple elements with the same key can be resolved by a `merge` function

For class member caches, the `merge` function will provide the logic for member overriding--when a method from earlier in the linearization (that is, further down the inheritance hierarchy) overrides a method from later in the linearization, `merge ~earlier ~later` will return `earlier`. If the method from higher up in the inheritance hierarchy should not be overridden by the lower (for instance, because it is concrete, and the lower/earlier method is abstract), `merge ~earlier ~later` will return `later`.

`Typing_class_member_cache.t` also accepts an `is_canonical` function, which allows it to stop advancing the sequence early if we encounter a member which we know cannot be replaced by any member from later in the linearization. For example, if we encounter a public, concrete, non-synthesized method, we know that no method from higher up in the inheritance hierarchy will modify or replace it in the final folded methods table, and we can return it immediately without examining the rest of the linearization.

Reviewed By: ljw1004

Differential Revision: D13568506

fbshipit-source-id: bb6d281f4bccd058285605562eb07f84702dff64