Improve documentation for fun(), inst_meth(), class_meth(), meth_caller()
Summary:
As part of a previous diff I noticed our documentation for `fun()` was incomplete.  It pointed to this page on HHVM which had little content:
* https://docs.hhvm.com/hack/reference/function/HH.fun/

I believe this was inadvertently changed by diff D2720738 which deleted the old content and pointed to the HHVM docs website.

I put back the original content and cleaned it up to document all the different types of legal use cases of each of the four functions:
* fun()
* inst_meth()
* class_meth()
* meth_caller()

Reviewed By: kmeht

Differential Revision: D13907996

fbshipit-source-id: ab9641d19cfd49a83b247ca9aa54648a3de56789