Add Explicit Support for Response Compression Shadowing
Summary:
As I work to improve compression of HTTP responses, it's useful to be able to
experiment with experimental compression schemes. Previously, I had attempted
to do this hackily, by adding experimental `ResponseCompressor`s that
reported themselves as both accepted and enabled, but then returned
`nullptr`.

It turned out that this introduced a subtle bug, in which the
`ResponseCompressorManager` would then fail to set `Vary: Accept-Encoding` on
the response if the request did not include any supported codings in its
`Accept-Encoding` header (since the fake compressor reported itself as
enabled and accepted).

This diff adds another method that `ResponseCompressor`s can implement,
`shouldShadow()`. When this returns true, the `ResponseCompressorManager`
will also call `compressResponse()` on that compressor, but without any user-
visible side-effects.

Reviewed By: terrelln

Differential Revision: D14215211

fbshipit-source-id: ff286be195f97e18b38b96d263502136aaa2528a