Kill DefFunc (Runtime changes)
Summary:
This diff removes the ability to define functions outside of top level and out of classes, i.e.
```
if (true) {
  function f() {}
}
```
is not disallowed. Typechecker already disallowed this behavior and the runtime was only allowing to maintain PHP support. Since we no longer need to maintain PHP compatibility, we are removing this from the language.

Reviewed By: paulbiss, alexeyt

Differential Revision: D14307801

fbshipit-source-id: 5c05b68c3c2a6319a228d40b056b409fb5734bfb