Clean up handling of get_class in the emitter.
Summary:
emit_expression.ml had special code for emitting calls to get_class with no args (using the emit_get_class_no_args helper), but it just did a normal function call, so remove this special handling.

In closure_convert.ml convert_expr we were trying to constant-fold calls to get_class with no arguments. This is not necessary for correctness, and we're about to remove the ability to call get_class with no args so remove this.

This made the fallback_to_empty_string argument to inline_class_name_if_possible redundant (always true), so simplify it away.

This also left convert_id as the only caller of inline_class_name_if_possible, so inline it into the caller, which also lets us simplify the logic.

Finally, in convert_id if we see the `__CLASS__` constant in a trait, don't convert it to a call to get_class with no args. Leave it as a constant and add code to emit_id in emit_expression.ml to emit it as a Self; ClsRefName sequence.

Reviewed By: vassilmladenov

Differential Revision: D14328351

fbshipit-source-id: adf3755d0bdb25147a212d01a06a905e44157a02