Fix IsTypeL's handling of uninits
Summary:
IsTypeL decided on `constprop()` twice. It would `constprop()` either if the
local was not uninit, *or* if the check itself would not throw.

This diff changes that *or* to *and* and does it at the proper location that
eliminates various redundant checks. As a side-effect, it allows constprop()
on `is_object()` in `IsTypeC`.

Reviewed By: markw65

Differential Revision: D14390760

fbshipit-source-id: 3c563ed7863e433cd8cde93d1e40c2c8177f9622