Use ExnNodeIds rather than ExnNode pointers
Summary:
I want to try to share identical php::Blocks between cloned functions (we
get a lot of clones due to trait flattening), but the ExnNode* in the
block makes that impossible.

This eliminates all stored ExnNode pointers, and replaces them with
ExnNodeIds. Since we no longer need address stability, we can allocate
the nodes themselves in a single CompactVector, rather than a
CompactVector of std::unique_ptrs (note that for the most part, we preserve
address stability anyway by pre-sizing the ExnNode vector).

This should make it easier to use copy_ptr to share Blocks.

Reviewed By: ricklavoie

Differential Revision: D14313945

fbshipit-source-id: 3dda8d6349690fb961f4c380312c6a209eaeeb1a