Use BlockIds in rpoBlocks
Summary:
When we switch to copy_ptr, mutating a block could modify its pointer,
so we don't want to store long lived pointers in the FuncAnalysis object.

Using block ids solves that problem, and also reduces the memory
requirements. It also turns out that a large number of users of
rpoBlocks only wanted the blockid anyway.

Reviewed By: ricklavoie

Differential Revision: D14335998

fbshipit-source-id: 67f22cf9139d2df42c8f32a808cb21b8d4efe9c9