Avoid mutating Bytecodes while emitting
Summary:
Some fields of some bytecodes, such as ClassIds and FuncIds need to be
translated at emission time.

We were doing that by mutating the bytecodes in place immediately
before emitting them. Once we start using copy_ptrs, its possible for
the Block containing such bytecodes to be shared - so two threads
could be mutating them at the same time.

One solution would be to pass a copy of each bytecode to emitter, and
mutate the copy. But copying bytecodes isn't free, so instead copy
just the ones that (might) need mutations.

Reviewed By: ricklavoie

Differential Revision: D14366310

fbshipit-source-id: 7fa2722db72763c633aae7c6891d7350cba178aa