Do not consider values of type XHPChild to be always truthy
Summary:
Since the XHPChild interface is implemented by Hack arrays and Stringish, values of type XHPChild may be falsy (`''`, `'0'`, empty Hack arrays). Because the truthiness utility functions do not currently account for this, Hack incorrectly reports that values of type XHPChild are always truthy.

This change handles XHPChild in the same way as Stringish--values of exactly the type XHPChild are reported as possibly falsy, and truthiness tests on them are considered sketchy (e.g., are you sure you want different behavior for `'0'` and `<x:frag>0</x:frag>`?).

Reviewed By: DavidSnider

Differential Revision: D14355018

fbshipit-source-id: 4300b5591d74bbb9a77522d3dd179890c2a88072