add setEncryptionIndex and setEncryptionName to ZipArchive
Summary: Currently there's no way to create an "encrypted" zip, which causes issues for some users. For example, it's a standard in security when handling malware to distribute malware files as zips with a well known password, to prevent antiviruses picking up the malware. However, encryption in zip is pretty rubbish so I've added a big warning in the docs.

Reviewed By: Ngo-The-Trung

Differential Revision: D14119412

fbshipit-source-id: 0757bf6cdd7b8487be69c6d2eee17838f4cd9bc4