Use copy_ptr for php::Blocks
Summary:
This allows us to share blocks, which happens a lot for Trait methods.

This can potentially save a lot of memory during the analysis phase of
the build, but once we start optimizing, the copy_ptrs will
copy-on-mutate, so this alone won't reduce peak memory.

I have a couple of ideas for solving that.

 - use a hash table to keep an interned set of blocks, or

 - emit the blocks right after optimizing them (and throw away
   php::Func).

I'll try some experiments on top of this.

Reviewed By: ricklavoie

Differential Revision: D14336089

fbshipit-source-id: 2281721b2bb6a6033b4951ec2b472935d417fd24