Add a fastpath for RepoOptions
Summary:
Use a thread local to cache the last config used and try to reuse that one, failing that scan through the hash table before actually looking at the filesystem to avoid making access() calls.

This isn't fantastic as it means we're likely to miss if new .hhvmconfig.hdf files are added, but since having multiple such files in one repo is a fatal anyway and we plan to overhaul this to use watchman eventually that seems acceptable.

Reviewed By: alexeyt

Differential Revision: D14255968

fbshipit-source-id: bae153c6e56db12f49976a5b0af36f401556c010