Resolve more type-constraints within HHBBC
Summary:
Type-constraint resolution within HHBBC worked by resolving the type-constraint
to a particular type, and then using that type to infer a better
AnnotType. However a function to map a type name to a particular AnnotType/class
already existed, so just use that instead. This has two advantages. First, it
allows us to resolve type-constraints to mixed to mixed directly. Second, if the
type-constraint resolves an object, we can mark it as such.

This lead to more cases of a pre-existing bug where if we resolved a
non-nullable type-hint to a nullable type, we'd display the original name with
?. To fix this, add a new flag to TypeConstraint to indicate whether the name
should be displayed with a ?, regardless of how it was resolved.

Reviewed By: jano

Differential Revision: D14410044

fbshipit-source-id: 929aa0fadb3aca560e705712651e88b6cdfc89cf