Add a simple pass to delay DecRefs to expose more opportunities to refcount opts
Summary:
This pass delays DecRef* instructions to try to expose more opportunities for
eliminating IncRef/DecRef pairs. The idea is to remember the last IncRef tX that
was seen and move DecRefs of other SSATmps later, until after a DecRef tX is
found.  If any other instruction that is not "irrelevant" is found along the
way, the set of DecRefs that are candidates to be moved has to be cleared.  For
now, this pass is only applied locally, within a basic block, as the
opportunities I've seen are all within a block (caused by the code sequence
emitted for storing a value to memory).

This diff also makes the identification of "irrelevant instructions" for
refcount opts more aggressive.  Instead of requiring all the alias sets for
instructions with GeneralEffects to be empty, we now only require them to not
touch heap locations.  The primary motivation for this was to allow
FinishMemberOp to be considered irrelevant (as it only affects the
member-instructions state).  I'm including this change in the same diff as it
alone appears to cause a mysterious, small regression.

Reviewed By: mofarrell

Differential Revision: D14119219

fbshipit-source-id: baba692bcd8008bd6aba1336035146f0b057c3a1