Add pos to AAST.stmt
Summary:
We currently have position info on statements on the legacy AST but not on the AAST. For compiling off the AAST, we want to add `pos` to `AAST.stmt` to ensure there are no differences in `srclocs` in the bytecode.

Position info on statements is also useful for typechecker errors.

Reviewed By: andrewjkennedy, losvald, Wilfred

Differential Revision: D13986819

fbshipit-source-id: e331ad5b74084f343831212a309f3a7a2eb04480