Add hh.conf option to require saved-state
Summary:
This diff supports a new option in hh.conf, "require_mini_state" (default false).

The intent: if "use_mini_state" is true, and if our attempt at loading saved-state failed, then we'll either fallback to a full init (if require_mini_state is false), or fail immediately (if require_mini_state is true).

This diff merely adds parsing for the option. It doesn't yet implement the fail-immediate behavior in the server, nor recovery in the client. That's left for a future diff.

Reviewed By: 2BitSalute

Differential Revision: D13541382

fbshipit-source-id: 73c78c3b58958b0a84cdf9aad5ab4cd31ff99945