ServerInit's fail-fast is sent to client
Summary:
The client has a stdio pipe to the monitor at first ("monitor time") over which they communicate messages. The monitor then hands off to the server. The server spends the first length of time doing init. After the init has done, the server acknowledges the client and sets up a two-way communication with it ("server time").

At monitor time, the client can asynchronously query the monitor about the server's status by (1) the server sends messages to the monitor updating its memory of the server's most recent status, (2) the client sends an RPC to the monitor to read that most recent status.

At server time, the server can send its progress messages directly to the client over stdio.

If the server terminates before handoff from the monitor, then the monitor itself dies and the stdio pipe gets EOF. If the server terminates after handoff, then the stdio pipe also gets EOF. But we have no mechanism for getting further details about the reasons behind that EOF.

Flow does have a mechanism. If the monitor/server is about to die for known controlled reasons, then they first synchronously send an "I'm about to die" message over stdio which the client receives. That way, on the subequent EOF, the client will know why.

It's important to know more details. That's so that the client can display a human-readable message explaining the cause of the death, and so it can also make an appropriate decision about whether to automaticaly restart the server.

Hack doesn't have any mechanism.

In the case of clientLsp, it always simply presents the message "hh_server is stopped". It also needs to gather 1 further bit of information, for whether it should automatically re-attempt a connection or not. This bit is communicated by the presence of the monitor: if the monitor remains alive then clientLsp will automatically reconnect; if the monitor is dead then clientLsp won't automatically reconnect.

That's a fragile and irritating mechanism. This diff introduces a second mechanism. We can't use the Flow approach because (for failures that happen after handoff but before the server has yet acknowledged the client) the server doesn't even have an stdio pipe with the client over which to communicate. We could change that, but it'd be a major change which I don't want to do now.

This diff introduces the following new mechanism.

When the monitor forms its connection to the server, it figures out the filename for a "finale file" specific to this server process. It uses "<PID>.fin" in the hack temp directory. It communicates this file's path to the client at the moment it hands off.

If there is a particular whitelisted reason for server death, then the server is able to write information to this file. So far the only whitelisted reason I've implemented is "server dies due to saved-state not loading but hh.conf said it was required".

The information it writes for now is the server exit code, a human-readable message, and a callstack.

The client could notionally read this file at any time after handoff. So far, the only occasion I've had to read and act upon it is after handoff and before "hello" (which gets sent once the server has finished its initialization and is ready to handle client requests).

Reviewed By: dabek

Differential Revision: D13823292

fbshipit-source-id: 8fac643230d3d76b9f724d94fa751140b0b4d08f