Remove codegen flag from lexer/parser
Summary:
I want to clean up some of the options we have in the lexer/parser now that we do not need to support PHP. The first of these is the usage of the codegen flag. For now I remove it from the lexer and parser. There is still a usage of the codegen flag as part of the smart constructor to remove trailing trivial, I assume for perf reasons. I'll leave it for now and clean it up later once I can determine it won't regress performance.

Particularly this seems to have disabled case insenitive keywords, but I have not removed that capability completely in this diff, since it also is tied to whether or a file is considered a Hack file or not.

Reviewed By: oulgen

Differential Revision: D14442511

fbshipit-source-id: f0893cc5f36375ac2dc1882670e29304169a7554