Add parameter and return type verification to reified primitives
Summary:
This diff adds parameter and return type verification to reified primitives as
```
function f<reify Ta, reify Tb>(Ta $a): Tb { return $b }
```
for this function we will check that `$a` is of type `Ta` and `$b` is of type `Tb`.
Similar to how regular verification works
```
function f <reify T>((T, T) $x) {...}
```
will only check that `$x` is a tuple and will not check the inner type. We can revisit this decision but for the time being we do not want to surprise developers with extra checks for consistency reasons.

Reviewed By: markw65

Differential Revision: D14099359

fbshipit-source-id: aa36d3d6ed64707fc26f85118a889e0a32412d58