Add ability to dump the reasons why the JIT decided to inline functions
Summary:
This also includes the computed inlining cost and its components.  This can be
useful when investigating differences in inlining decisions in A/B comparisons.

The previous DumpInlRefuse option is renamed to DumpInlDecision and turned into
an integer.  When passed 1, only refusals are logged; when passed 2, both
refusals and accepts are logged.

Reviewed By: binliu19

Differential Revision: D14456487

fbshipit-source-id: a6f970d5436f9380e7f4bfc59cbeef1d12f66ac1