Remove a redundant String::operator+
Summary:
I had some code like this:

```
if (!m_root.empty() && path.get()->data()[0] != '/' ) {
  path = m_root + path;
}
```

I assumed that checking `m_root` for emptiness probably didn't matter, and I thought that we could save a copy by moving `path`, so I changed the code to

```
if (path.get()->data()[0] != '/') {
  path = m_root + std::move(path);
}
```

This caused a segfault. The segfault stops when we remove the check for emptiness, or remove the `std::move`. It turns out that most implementations of `String::operator+` special-case empty strings, but

```
friend String operator+(const String & lhs, String&& rhs);
```

forgot to check for emptiness. We don't actually get any efficiency out of the move in this method, so let's lower the surface area a bit and remove this method.

Reviewed By: markw65

Differential Revision: D14491914

fbshipit-source-id: 2a7bdd0a371afa9aa0bc31599d11fa82ebd01bb7