Fix inline analysis dependencies
Summary:
When doing lookup_foldable_return_type, we limit the inline nesting
depth to 2. This can cause inconsistent results for something like:

  function a() { return b("b"); }
  function b($x) { return $x . c("c"); }
  function c($x) { return $x . d(); }
  function d() { return "d"; }

On the first round, we don't know that d is effect free, and returns a
litteral, so when doing inline analysis of b into a, we'll give up,
because we hit the depth limit.

In the same round, we'll successfully flatten d() into c, and c into b
(because neither hits the depth limit), but there's nothing to trigger
re-analysis of a.

This adds an InlineDepthLimit dependency, and applies it between a and
d so that when d resolves to an effect_free function, we re-analyze a.

Reviewed By: ricklavoie

Differential Revision: D14497911

fbshipit-source-id: 003447a6013992b156514293ed3167a91d631bf3