Fix segfault when calling StringData::Make with an empty StringPiece
Summary:
I had some code like this:

```
if (!m_root.empty() && path.get()->data()[0] != '/' ) {
  path = m_root + path;
}
```

I assumed that checking `m_root` for emptiness probably didn't matter, and I thought that we could save a copy by moving `path`, so I changed the code to

```
if (path.get()->data()[0] != '/') {
  path = m_root + std::move(path);
}
```

This caused a segfault. The segfault stops when we remove the check for emptiness, or remove the `std::move`. It turns out that most implementations of `String::operator+` special-case empty strings, but

```
friend String operator+(const String & lhs, String&& rhs);
```

doesn't. Note that if a `String` is empty and its `StringData*` is `nullptr`, then `String::slice` will return a default-constructed `folly::StringPiece`.

But we should be able to pass a default-constructed `StringPiece` to `StringData::Make` without segfaulting... I don't know why `memcpy` doesn't check if the size passed to it is 0. But given that it seems not to check its size, let's check that the argument passed to it isn't `nullptr` ourselves.

Reviewed By: markw65

Differential Revision: D14492459

fbshipit-source-id: ad2ed95a2a775ff2ef48f663d24be7d6fcb17efc