Simplify DCE's isCFPushTaken() logic
Summary:
The implementation of isCFPushTaken() is trying to be generic, but these rules
have nothing common with whether the opcode can jump to a target without
pushing anything.

For example, this is wrong for YieldFromDelegate, which should be
instrIsNonCallControlFlow() (being fixed by a diff on top of this one), as it
is doing control flow outside of Yield semantics, but YieldFromDelegate pushes
a cell even if it jumps.

Let's explicitly list all 2 opcodes (one of them was already special handled)
for which this function was able to return true, i.e. opcodes that can jump to
a target without pushing anything.

Reviewed By: ricklavoie

Differential Revision: D14531993

fbshipit-source-id: 9750939a8566f121da1b91995604574902bfa9a3