bug in tyvar_is_solved
Summary:
See the added test for a regression test.
`tyvar_is_solved` use to check that a type variable was bound to a concrete type in tenv. However now some type variables are solved by being bound to other type variables, so we simply check that the type variable expands to something different than itself. Failing to do that would readd an already solved type variable with no bound and no variance in `solve_tyvar`, which would then be solved to `nothing`.

Reviewed By: andrewjkennedy

Differential Revision: D14454694

fbshipit-source-id: 271d23e5dfa37a34d8b0405704ec83598b267929