fix regression in completeness due to removing remove_equivalent_tyvars
Summary:
See the added test for how the regression was caused. Basically, we now have constraints sets like:
```
C, #1 <: #2 <: #1
C, #2 <: #1 <: #2
```
so when `expand_type_and_solve`ing #2, we would fail to solve it to C.
Solution in this diff: `solve_tyvar` now recursively solve until we get a concrete type.

Reviewed By: andrewjkennedy

Differential Revision: D14438531

fbshipit-source-id: aee3ca20c7aedd8197aaa87924712af5b9d33b7b