expand_type_and_solve should normalize unions
Summary:
Sometimes we have type variables appearing under Tunresolved or Tunion. When we expand_type_and_solve on them, we might miss a possible simplification of the union that we could have done after solving the variable. For example, if $x is a (#1 | vec<nothing>), appending to $x will do expand_type_and_solve on the type of $x, see that it is a union, try to append on each of the elements of the union, early solve #1 to nothing and fail.

In this diff, expand_type_and_solve is a call to Union.simplify_unions, which normalizes any top-level unions, and is made to solve any type variable occuring under the union using an extra ~on_tyvar argument.

Reviewed By: andrewjkennedy

Differential Revision: D14505691

fbshipit-source-id: 1456b0166fc86a24015d310a6a33ccc7add1793c