account for return type of lambda when setting variance
Summary:
When checking lambda's passed as arguments of functions, they get an expected type, which, if the function is generic, contains type variables, for example `function (Cov<#1>, Contrav<#2>): #1`.

So far, before typechecking the body of the lambda, we set the variance of the type variables according to their variance in the types of the arguments, so here #1 would be covariant and #2 contravariant. However we also need to account for the variance of those type variables in the return type, but flipped, meaning here #1 would also be contravariant.

Why is that? Consider a 'reduce' function:
```
function reduce<Ta, Tv>(Ta $acc, Tv $v, (function (Ta, Tv): Ta) $f): Ta
```
It's currified version is
```
function reduce<Ta, Tv>(Ta $acc, Tv $v): function(function (Ta, Tv): Ta): Ta
```
and in the currified function, if #1 is for Ta and #2 for Tv, #1 appears both covariantly and contravariantly in the return type of reduce because of its appearance in the return type of the lambda parameter of reduce.

Reviewed By: andrewjkennedy

Differential Revision: D14504272

fbshipit-source-id: 158018fddab9e8941c226b172f7227ebc6d89009