Fix equivLoc when a stack base is updated
Summary:
If we updated the type of a stack base, we would also set its equivLoc
to 0.

I don't think this could currently be a problem with generated
bytecode, because we always pop the base in the final op, without
having used it for any other purpose.

Its at least theoretically possible though to leave the base on the
stack after the minstr sequence is done, and eg compare it against the
first local, which hhbbc could then optimize away...

Reviewed By: ricklavoie

Differential Revision: D14540040

fbshipit-source-id: 79ab49c23fb218dd525a9e748562d43f70edc0f4