Treat error-suppression comments like hackfmt-ignore comments
Summary:
We should not format nodes following an error-suppression comment, since we might move the suppressed error to a new line.

Because of the hackfmt-ignore feature, we already support leaving the formatting of a node unchanged when it has a particular leading comment. We just need to also trigger this behavior for error-suppression comments.

Reviewed By: 2BitSalute

Differential Revision: D14524402

fbshipit-source-id: 32d13cadb747cd6f0b81a51ccf9fd03f7b0254ba