New inference: check index type in array get during TAST checks
Summary: During check of array access expressions such as `e1[e2]`, we check that the type of `e2` is compatible with the key type derived from the type of the array (or collection) `e1`. This is strictly speaking not a "type system" check, because any array is a subtype of `KeyedContainer<mixed,t>` for some `t`, and so `e2` is always valid. But typically we have a more precise type for `e1` so it's useful to know, for example, that a string index can't be used to lookup in a `dict<int,bool>`. However, under `new-inference`, we don't always have precise information at the point of checking the array lookup. Instead, we can defer the check until TAST checking time. This is only necessary when the key type is present in the type of the collection itself: for vectors (and vector-like types), we know that the key should be an `int`, and can check it as normal during type inference.

Reviewed By: vsiles

Differential Revision: D14520917

fbshipit-source-id: 4e38206966a358c4a80eab713ca8386788dbf910