Remove HH_FIXMEs loaded from the saved state for any changed files before re-parsing them
Summary: We store HH_FIXMEs in two shared memory heaps in the Fixmes module, HH_FIXMES and DECL_HH_FIXMES. These heaps are populated whenever we parse a file, and are intended to be cleared just before we reparse a file. The clearing happens in ServerTypeCheck, which was historically the single point where we would reparse files and needed to guarantee this invariant. However, now that we load HH_FIXMEs from the saved state, we now need to remove any HH_FIXMEs for changed files from shared memory before re-parsing them.

Reviewed By: 2BitSalute

Differential Revision: D14574327

fbshipit-source-id: b0820e9b77eae8c4ccd5a3239262e5f20f0890a1