Allow hugifyText() placement before hot text
Summary:
Check if hugifyText() start does not fall into hot text. The more
accurate check involves the knowledge of the size of hugifyText()
which we don't have at compile-time. The borderline case is when
hugifyText() starts before 'from' and ends after it resulting
in the runtime crash. Arguably, this is preferable to having
a more conservative check that fails to hugify the legitimate
placement of hugifyText() right before the hot text causing a
silent performance regression.

Reviewed By: rafaelauler

Differential Revision: D14565848

fbshipit-source-id: 1989ba30e330b172a4289462aa1816f33882cec4