Fix `hh_single_type_check --decl-compare`
Summary:
This tests incremental mode; we do this by doing trivial changes to
the file, originally prepending a line; this got us:

```

<?hh
```

... with a markup section. This was then changed to append it after the first
line, assuming it was a header line, and add it after. If there's a shebang
line, we end up with:

```

<?hh
```

In both of these cases, we have a markup section containing an empty line. Instead, add the newline after any shebang and/or header

Differential Revision: D14601254

fbshipit-source-id: 471259c311d531e0426534ee7a59b54039cce3b4