add hhjs repo root config
Summary: Add HHJSRepoRoot flag because the file path sent to hackc is absolute and we don't want to include the absolute path in  the autoload_map. We use the root property in a later diff to strip the machine specific info: ie data__user__cfernand__www/ from the function path.

Reviewed By: gkz

Differential Revision: D14555730

fbshipit-source-id: 042d0963cf3b6ebccd962bc7f56d3e6062619830