Check dumpTCAnnotation() when dumping inlining decisions
Summary:
dumpTCAnnotation() tests a few conditions to determined whether TC annotations
should be dumped for a given function and translation kind, but the result of
this function was not being checked when dumping inlining decisions.  This diff
fixes that.  Unfortunately, a number of places where the inlining decisions are
being logged don't have the caller Func and the translation kind that
dumpTCAnnotation takes as arguments.  Instead of threading these arguments in a
bunch of places, this diff changes some arguments from references to Annotations
into pointers, with the meaning that, if the pointer is non-null, then inlining
decisions should be dumped.

Reviewed By: binliu19

Differential Revision: D14588334

fbshipit-source-id: effe7d5efb0040c5ffcd7444622014c942b608b3